LPS-89711 regen

+++ modules/apps/headless/headless-document-library/headless-document-library-api/src/main/java/com/liferay/headless/document/library/resource/DocumentResource.java
@@ -45,7 +45,7 @@ public interface DocumentResource {
 	@Path("/document/{id}")
 	@Produces({"*/*"})
 	@RequiresScope("headless-document-library-application.read")
-	public Document getDocument(@PathParam("id") Integer id) throws Exception;
+	public Document getDocument(@PathParam("id") Long id) throws Exception;
 
 	@GET
 	@Path("/documents-repository/{parent-id}/document")
