+++ modules/apps/headless/headless-document-library/headless-document-library-impl/src/main/java/com/liferay/headless/document/library/internal/resource/BaseDocumentResourceImpl.java
@@ -31,7 +31,7 @@ import javax.annotation.Generated;
 public abstract class BaseDocumentResourceImpl implements DocumentResource {
 
 	@Override
-	public Document getDocument(Integer id) throws Exception {
+	public Document getDocument(Long id) throws Exception {
 		return new Document();
 	}
